requirement.txt
fastapi
uvicorn
python-multipart
motor
pydantic-settings
tensorflow
ultralytics
torch
torchvision
xgboost
scikit-learn
pillow
numpy
opencv-python
razorpay
python-jose[cryptography]
passlib[bcrypt]
langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transform
groq
python-dotenv
tqdm
sqlalchemy
pymysql
alembic
